\name{BusRouteCreelSurvey-package}
\alias{BusRouteCreelSurvey-package}
\alias{BusRouteCreelSurvey}
\docType{package}
\title{Simulate a Bus Route Creel Survey}
\description{Create an angler population, sample the population with a user-specified suvey times, and calculate metrics from a bus route-type creel survey.}
\details{
\tabular{ll}{Package: \tab BusRouteCreelSurvey\cr
Version: \tab 0.2.1\cr
Date: \tab 2015-01-02\cr
Title: \tab Simulate a Bus Route Creel Survey\cr
Author: \tab Steven Ranney\cr
Maintainer: \tab Steven H. Ranney <steven.ranney@gmail.com>\cr
Depends: \tab R (>= 3.0.2)\cr
License: \tab GPL-3\cr}
}
\author{Steven Ranney}

\keyword{ package }


